// apps/web/lib/api/types/MarketPositions.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { HTTPValidationError } from "./HTTPValidationError"
import type { PagePositionView } from "./PagePositionView"

export const marketPositionsQueryParamsSortEnum = {
  value: "value",
  gain: "gain",
  gain_percent: "gain_percent",
  cost: "cost",
  quantity: "quantity",
  name: "name",
} as const

export type MarketPositionsQueryParamsSortEnumKey =
  (typeof marketPositionsQueryParamsSortEnum)[keyof typeof marketPositionsQueryParamsSortEnum]

export const marketPositionsQueryParamsDirectionEnum = {
  asc: "asc",
  desc: "desc",
} as const

export type MarketPositionsQueryParamsDirectionEnumKey =
  (typeof marketPositionsQueryParamsDirectionEnum)[keyof typeof marketPositionsQueryParamsDirectionEnum]

export type MarketPositionsQueryParams = {
  /**
   * @default "value"
   * @type string | undefined
   */
  sort?: MarketPositionsQueryParamsSortEnumKey
  /**
   * @default "desc"
   * @type string | undefined
   */
  direction?: MarketPositionsQueryParamsDirectionEnumKey
  /**
   * @minLength 1
   * @maxLength 100
   * @default 20
   * @type integer | undefined
   */
  limit?: number
  /**
   * @minLength 0
   * @default 0
   * @type integer | undefined
   */
  offset?: number
}

/**
 * Page[PositionView]
 * @description Successful Response
 */
export type MarketPositions200 = PagePositionView

/**
 * HTTPValidationError
 * @description Validation Error
 */
export type MarketPositions422 = HTTPValidationError

export type MarketPositionsQueryResponse = MarketPositions200

export type MarketPositionsQuery = {
  Response: MarketPositions200
  QueryParams: MarketPositionsQueryParams
  Errors: MarketPositions422
}